move some key files to brand directory
