+++ packages/app/src/brand/logo/MainLogo.tsx
