+++ packages/app/src/brand/logo/SidePanelLogo.tsx
